Move mc_manager onto migration tooling
