+++ mc_manager/api/BUILD.bazel
@@ -2,15 +2,6 @@ load("@py_deps//:requirements.bzl", "requirement")
 load("//tools/build_rules:api_image.bzl", "api_image")
 
 py_library(
-    name = "config_py",
-    srcs = ["config.py"],
-    visibility = ["//mc_manager:__subpackages__"],
-    deps = [
-        "//base:flask_app_py",
-    ],
-)
-
-py_library(
     name = "api_lib",
     srcs = glob(["**/*.py"]),
     imports = [".."],
