+++ mc_manager/api/gql/types/server_backup.py
@@ -14,8 +14,8 @@ from graphql import (
 )
 from sqlalchemy import desc
 
-from mc_manager.api.config import db
 from mc_manager.api.models.server_backup import ServerBackup
+from mc_manager.config import db
 
 serverBackupStateEnum = GraphQLEnumType(
     "ServerBackupState",
