+++ mc_manager/api/gql/types/server_log.py
@@ -14,9 +14,9 @@ from graphql import (
 )
 from sqlalchemy import desc
 
-from mc_manager.api.config import db
 from mc_manager.api.models.server_backup import ServerBackup
 from mc_manager.api.models.server_log import ServerLog
+from mc_manager.config import db
 
 serverLogStateEnum = GraphQLEnumType(
     "ServerLogState",
