+++ mc_manager/api/models/server_backup.py
@@ -1,7 +1,7 @@
 import datetime
 
-from mc_manager.api.config import db
 from mc_manager.api.models.server import Server
+from mc_manager.config import db
 
 
 class ServerBackup(db.Model):
