+++ mc_manager/api/models/server_log.py
@@ -1,8 +1,8 @@
 import datetime
 
-from mc_manager.api.config import db
 from mc_manager.api.models.server import Server
 from mc_manager.api.models.server_backup import ServerBackup
+from mc_manager.config import db
 
 
 class ServerLog(db.Model):
