+++ mc_manager/docker-compose.yaml
@@ -9,7 +9,7 @@ services:
       - pg
   migration:
     <<: *api
-    command: [ "/mc_manager/api/migrations/binary", "upgrade"]
+    command: [ "/mc_manager/api/migrations/binary", "/mc_manager/api/migrations", "./binary.runfiles/_main/mc_manager/api/migrations/alembic.ini", "upgrade"]
     restart: no
   worker:
     <<: *api
